chore: Remove ConnectionFlowRepositoryInterface and ConnectionFlowRepository
This commit removes the ConnectionFlowRepositoryInterface and ConnectionFlowRepository files, as they are no longer needed in the codebase. These files were deleted as part of a code cleanup process to remove unused or redundant code. This change does not affect the functionality of the application.

Please note that this commit message does not include any meta information such as issue references, tags, or author names.