feat: Update models and configurations
This commit updates the models and configurations in the codebase. It adds the 'description' field to the 'Client', 'Flow', and 'Connection' models. It also updates the 'Currency' model to include the 'rate' and 'code' fields. Additionally, the 'evolution.php' configuration file is modified to use the 'evolution_api:8080' URL. Finally, the 'ConnectionFlow', 'ConnectionSetting', 'FlowSession', 'ObserverServiceProvider', 'CreateConnectionRequest', 'UserSeeder', 'MessengerServiceInterface', and 'Message' models are also modified.