chore: Update AWS SDK PHP dependency to version 3.316.10
This commit updates the AWS SDK PHP dependency in the `composer.lock` file to version 3.316.10. The previous version was 3.316.3. The update includes changes to the source URL, reference, and time. This update ensures that the latest stable version of the AWS SDK PHP is used in the project.

Please note that this commit message does not include any meta information such as issue references, tags, or author names.