refactor: Update local_domain in mail.php to handle default APP_URL
This commit updates the `local_domain` configuration in the `mail.php` file to handle the default `APP_URL` value. Previously, the `local_domain` was set to the host of the `APP_URL` environment variable. However, if the `APP_URL` was not set, it would throw an error. This change ensures that the `local_domain` falls back to `http://localhost` when `APP_URL` is not defined. This improves the reliability of the email functionality in the application.

Please note that this commit message does not include any meta information such as issue references, tags, or author names.